Added writing the addon file to realtime console

+++ SharpGMad/RealtimeCommandline.cs
@@ -9,13 +9,18 @@ namespace SharpGMad
     partial class Program
     {
         static UpdatableAddon addon;
+        static string filePath;
         static string CommandlinePrefix = "SharpGMad>";
 
-        static int RealtimeCommandline()
+        static int RealtimeCommandline(string strFile)
         {
             Console.ForegroundColor = ConsoleColor.Green;
             Console.WriteLine("Realtime command-line activated.");
             Console.ResetColor();
+
+            // Autoload addon from command line arguments
+            if (strFile != String.Empty && strFile != null)
+                LoadAddon(strFile);
 
             while (true)
             {
@@ -26,6 +31,12 @@ namespace SharpGMad
                 switch (args[0])
                 {
                     case "load":
+                        if (addon is UpdatableAddon)
+                        {
+                            Output.Warning("An addon is already open. Please close it first.");
+                            break;
+                        }
+
                         try
                         {
                             LoadAddon(args[1]);
@@ -38,9 +49,21 @@ namespace SharpGMad
 
                         break;
                     case "list":
+                        if ( addon == null )
+                        {
+                            Output.Warning("No addon is open.");
+                            break;
+                        }
+
                         ListFiles();
                         break;
                     case "remove":
+                        if ( addon == null)
+                        {
+                            Output.Warning("No addon is open.");
+                            break;
+                        }
+
                         try
                         {
                             RemoveFile(args[1]);
@@ -52,8 +75,45 @@ namespace SharpGMad
                         }
 
                         break;
+                    case "abort":
+                        CloseAddon();
+                        break;
+                    case "push":
+                        Push();
+                        break;
+                    case "close":
+                        Push();
+                        CloseAddon();
+                        break;
+                    case "?":
+                    case "help":
+                        Console.WriteLine("Available commands:");
+                        Console.WriteLine("load <filename>                    Loads <filename> addon into the memory");
+
+                        if (addon is UpdatableAddon)
+                        {
+                            Console.WriteLine("list                               Lists the files in the memory");
+                            Console.WriteLine("remove <filename>                  Removes <filename> from the archive");
+                            Console.WriteLine("push                               Writes the addon to the disk");
+                            Console.WriteLine("close                              Writes addon and closes it");
+                            Console.WriteLine("abort                              Unloads the addon from memory, dropping all changes");
+                        }
+
+                        Console.WriteLine("help                               Show the list of available commands");
+                        Console.WriteLine("exit                               Exits");
+
+                        break;
+                    case "exit":
+                        if (addon is UpdatableAddon)
+                        {
+                            Output.Warning("Cannot exit. An addon is open!");
+                            break;
+                        }
+
+                        return 0;
+                        //break;
                     default:
-                        Console.WriteLine("Unknown operation.");
+                        Output.Warning("Unknown operation.");
                         break;
                 }
             }
@@ -64,22 +124,11 @@ namespace SharpGMad
             Console.WriteLine("Loading file...");
 
             addon = new UpdatableAddon();
+            filePath = filename;
+            CommandlinePrefix = filename + ">";
 
             Reader r = new Reader(filename);
-
-            /*Addon.Reader r = new Addon.Reader();
-            if (!r.ReadFromFile(filename))
-            {
-                Console.WriteLine("There was a problem opening the file.");
-                return;
-            }
-
-            if (!r.Parse())
-            {
-                Console.WriteLine("There was a problem parsin the file.");
-                return;
-            }*/
-            
+                        
             addon.Author = r.Author;
             addon.Title = r.Name;
             addon.Description = r.Description;
@@ -107,22 +156,6 @@ namespace SharpGMad
             addon.UpdateInternalStream();
 
             Console.WriteLine("Addon opened successfully.");
-
-            FileStream filestream;
-            try
-            {
-                filestream = new FileStream(filename, FileMode.Open, FileAccess.ReadWrite, FileShare.Read);
-            }
-            catch (Exception e)
-            {
-                Console.WriteLine("There was a problem opening the file.");
-                Console.WriteLine(e.Message);
-                return;
-            }
-
-            StreamDiffer sd = new StreamDiffer(filestream);
-            sd.Write(addon.Buffer);
-            sd.Push();
         }
 
         static void ListFiles()
@@ -137,6 +170,46 @@ namespace SharpGMad
         static void RemoveFile(string filename)
         {
             addon.RemoveFile(filename);
+            addon.UpdateInternalStream();
+        }
+
+        static void Push()
+        {
+            if (addon == null)
+            {
+                Output.Warning("No addon is open.");
+                return;
+            }
+
+            FileStream fileStream;
+            try
+            {
+                fileStream = new FileStream(filePath, FileMode.Open, FileAccess.ReadWrite, FileShare.Read);
+            }
+            catch (Exception e)
+            {
+                Output.Warning("There was a problem opening the file.");
+                Console.WriteLine(e.Message);
+                return;
+            }
+
+            StreamDiffer sd = new StreamDiffer(fileStream);
+            sd.Write(addon.Buffer);
+            sd.Push();
+
+            fileStream.Close();
+        }
+
+        static void CloseAddon()
+        {
+            if (addon == null)
+            {
+                Output.Warning("No addon is open.");
+                return;
+            }
+
+            addon = null;
+            CommandlinePrefix = "SharpGMad>";
         }
     }
 }
